HOLO-66 <chore: 안쓰는 컴포넌트 삭제>
