+++ src/app/mypage/[id]/_components/mylike/LikeCard.tsx
