+++ src/app/mypage/_components/MypageForm.tsx
@@ -1,8 +1,8 @@
 // src/app/mypage/_components/mypageform.tsx
 "use client";
 
-import MyCommentList from "@/app/mypage/[id]/_components/my-comment/useMyCommentList";
-import MyLikeList from "@/app/mypage/[id]/_components/mylike/useMyLikeList";
+import MyCommentList from "@/app/mypage/[id]/_components/my-comment/MyCommentList";
+import MyLikeList from "@/app/mypage/[id]/_components/mylike/LikeList";
 import MyWritingList from "@/app/mypage/[id]/_components/mypost/WritingList";
 import DeleteAccount from "@/app/mypage/_components/DeleteAccount";
 import UserProfile from "@/app/mypage/_components/UserProfile";
