+++ src/app/sign-in/_types/signInType.ts
@@ -1,18 +1,3 @@
-//주석단것 기능테스트해보고 지울예정
-
-// export type FormData = {
-//   email: string;
-//   nickname: string;
-//   password: string;
-//   checkPassword: string;
-// };
-
-// export type SignUpPayload = {
-//   email: string;
-//   password: string;
-//   nickname: string;
-// };
-
 
 export type SignInPayload ={
   email: string;
