--- src/app/sign-in/_types/userType.ts
@@ -1,9 +1,0 @@
-//이것도 기능테스트후 삭제예정
-
-// import { Tables } from "@/lib/types/supabase";
-
-
-// export type User = Tables<'users'> ;
-
-
-
