--- src/app/sign-up/_types/userType.ts
@@ -1,13 +1,0 @@
-//기능 테스트후 삭제예정
-
-
-// import { Tables } from "@/lib/types/supabase";
-
-
-// export type User = Tables<'users'> ;
-
-
-
-
-
-
